+++ internal/plugin/service/byoc/permissions/permissions_test.go
@@ -86,6 +86,12 @@ resource "aiven_byoc_permissions" "example" {
 				),
 			},
 			{
+				Config: updatedConfig,
+				ConfigPlanChecks: resource.ConfigPlanChecks{
+					PreApply: []plancheck.PlanCheck{plancheck.ExpectEmptyPlan()},
+				},
+			},
+			{
 				ResourceName:      resourceName,
 				ImportState:       true,
 				ImportStateVerify: true,